requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bitarray==2.1.3
bleach==4.1.0
blinker==1.4
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==37.0.4
cx-Oracle==8.2.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
elasticsearch==7.15.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.0
hdfs==2.6.0
idna==2.10
importlib-metadata==4.5.0
importlib-resources==5.4.0
impyla==0.17.0
interval==1.0.0
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
msoffcrypto-tool==5.0.0
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.20.3
olefile==0.46
openpyxl==3.0.7
packaging==21.3
panda==0.3.1
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
prettytable==3.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.19
psycopg2==2.9.3
ptpython==3.0.19
pure-sasl==0.6.2
pycparser==2.21
pyecharts==1.9.1
pygame==2.1.2
Pygments==2.9.0
PyHive==0.6.4
PyMySQL==1.0.2
pyparsing==3.0.7
pypinyin==0.46.0
pyrsistent==0.18.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pywin32==303
pywinpty==2.0.2
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rangedict==0.1.7
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
scipy==1.7.3
Send2Trash==1.8.0
simplejson==3.18.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.17
terminado==0.13.1
testpath==0.5.0
thrift==0.11.0
thrift-sasl==0.4.3
thriftpy2==0.4.14
tldextract==3.2.0
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
jydata-common @ file:///D:/workSpace/jiaoben/python/jydata_common/dist/jydata_common-1.0.0.dev0-py3-none-any.whl
jydata-flask @ file:///D:/workSpace/jiaoben/python/jydata-flask/dist/jydata_flask-1.0.0.dev0-py3-none-any.whl
